Allow using some privisioner flags in grouping.
Some flag values can be used to split groups. The canonical use case
here is the route53.SetID. We want to tret records with different
SetIDs as entirely different groups of records. This allows some
records to be controller outside of dubber, without dubber attempting
to remove them if they collide with records dubber itself is managing.