provisioner: owner groups
Allow a set of regexps that define, based on RR record comment tags,
the records that this dubber instance owns.

If all the configured regexps for a provisioner match a record, and
that record is present in the remote provisioner zone, but is not
wanted by the current discovery scrap, dubber will be allowed to
delete the record.